Significant optimization; new config option to restore old behavior
In previous versions, HSV recalculated everything about the judgment
popups every time the game did. And because the score for a cut isn't
finalized until shortly after the cut (to capture the follow-through
movement of the saber), this meant a fair amount of processing was being
done essentially every frame for a split second after each block was
cut. This caused performance issues for some users (#3).

This update changes that behavior. Now, the popup will update twice:
once when it's created, and once when the score is finalized. However,
this means you'll briefly see an inaccurate score before it jumps up to
the correct one, rather than seeing it update every frame and thus
approach the final score as the saber moves. A new config option,
"doIntermediateUpdates", has been added to restore the old behavior for
users with the performance to spare who want the most accurate possible
information as they cut notes. (This option is semi-recommended in
combination with the "useFixedPos" setting from last update, as:
- the jump is more noticeable when the judgments are in a fixed position
- since each new judgment replaces the previous one, they may not get a
  chance to update for the final score at all
- with only one judgment active at a time, the performance burden is
  smaller)

Note that if your config was modified from the default,
"doIntermediateUpdates" will be enabled, since it was technically the
default value in previous versions. If you've been experiencing
performance issues, disabling it should help significantly.